perf(bsseq): resize the matrix for every region instead of every CpG
Because the resizing was done in the matrix population step for each
CpG, it was resizing heavily with every CpG. This introduces an extra
loop, but resizes only once for every new set of CpGs in a RegionQuery
and then populates. So worst case should only be as many allocations as
regions.